requirements.txt
Django>=4.2
pandas>=2.0
matplotlib>=3.7
python-dotenv>=1.0
# For MySQL backends: either mysqlclient or mysql-connector-django
mysql-connector-python>=8.0
